+++ app/Service/Student/StudentListService.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace App\Service;
+namespace App\Service\Student;
 
 use App\Models\Resume;
 use App\Models\Tag;
